Learning and Hobbies/Br0GPT/requirements.txt
os
shutil
PyPDF2
flask
farm-haystack
openai-whisper
